Make GzipMiddleware BREACH resistant and add GzipMiddleware to the default middleware classes to avoid the use of BREACH vulnerable versions in the user's webserver
