+++ django/conf/global_settings.py
@@ -457,7 +457,7 @@ MIDDLEWARE_CLASSES = (
     'django.contrib.auth.middleware.AuthenticationMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
 #     'django.middleware.http.ConditionalGetMiddleware',
-#     'django.middleware.gzip.GZipMiddleware',
+    'django.middleware.gzip.GZipMiddleware',
 )
 
 ############
